Split Y-cable and Xcvrd tests into different files.
The Y-cable tests now reside in test_y_cable_helper.py
